pyproject.toml
[tool.poetry]
name = "my-rss-app"
version = "2.0.0"
description = "RSS aggregator with web server and persistent storage"
authors = ["Nils Tejedor <ntejedov@example.com>"]
readme = "README.md"
license = "MIT"
package-mode = false

[tool.poetry.dependencies]
python = "^3.9"
aiohttp = "^3.9.5"
beautifulsoup4 = "^4.12.3"
feedparser = "^6.0.11"
Flask = "^3.0.3"
gunicorn = "^22.0.0"
lxml = "^5.2.2"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
PyYAML = "^6.0.1"
requests = "^2.32.3"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.7.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
